refactor(deposit): move usecases into commands
