+++ src/features/deposit/commands/match-deposit.usecase.spec.ts
@@ -1,12 +1,12 @@
 import { MatchDepositUseCase } from './match-deposit.usecase';
-import { InMemoryProvisionalDonationRepository } from '../../infrastructure/repositories/in-memory-provisional-donation.repository';
-import { Deposit } from '../../domain/entities/deposit.entity';
-import { ProvisionalDonation } from '../../domain/entities/provisional-donation.entity';
+import { InMemoryProvisionalDonationRepository } from '../infrastructure/repositories/in-memory-provisional-donation.repository';
+import { Deposit } from '../domain/entities/deposit.entity';
+import { ProvisionalDonation } from '../domain/entities/provisional-donation.entity';
 import { EventEmitter2, EventEmitterModule } from '@nestjs/event-emitter';
-import { DepositMatchedEvent } from '../../domain/events/deposit-matched.event';
-import { DepositUnmatchedEvent } from '../../domain/events/deposit-unmatched.event';
-import { DepositPartiallyMatchedEvent } from '../../domain/events/deposit-partially-matched.event';
-import { GiftogetherExceptions } from '../../../../filters/giftogether-exception';
+import { DepositMatchedEvent } from '../domain/events/deposit-matched.event';
+import { DepositUnmatchedEvent } from '../domain/events/deposit-unmatched.event';
+import { DepositPartiallyMatchedEvent } from '../domain/events/deposit-partially-matched.event';
+import { GiftogetherExceptions } from '../../../filters/giftogether-exception';
 import { Test, TestingModule } from '@nestjs/testing';
 import { User } from 'src/entities/user.entity';
 import { ImageType } from 'src/enums/image-type.enum';
