+++ src/features/deposit/deposit.controller.spec.ts
@@ -1,8 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { DepositController } from './deposit.controller';
 import { DepositService } from './deposit.service';
-import { UploadDepositUseCase } from './application/usecases/upload-deposit.usecase';
-import { MatchDepositUseCase } from './application/usecases/match-deposit.usecase';
+import { UploadDepositUseCase } from './commands/upload-deposit.usecase';
+import { MatchDepositUseCase } from './commands/match-deposit.usecase';
 import { InMemoryDepositRepository } from './infrastructure/repositories/in-memory-deposit.repository';
 import { InMemoryProvisionalDonationRepository } from './infrastructure/repositories/in-memory-provisional-donation.repository';
 import { EventEmitterModule } from '@nestjs/event-emitter';
