+++ src/features/deposit/deposit.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { DepositController } from './deposit.controller';
 import { DepositService } from './deposit.service';
 import { EventEmitterModule } from '@nestjs/event-emitter';
-import { UploadDepositUseCase } from './application/usecases/upload-deposit.usecase';
-import { MatchDepositUseCase } from './application/usecases/match-deposit.usecase';
+import { UploadDepositUseCase } from './commands/upload-deposit.usecase';
+import { MatchDepositUseCase } from './commands/match-deposit.usecase';
 import { InMemoryDepositRepository } from './infrastructure/repositories/in-memory-deposit.repository';
 import { InMemoryProvisionalDonationRepository } from './infrastructure/repositories/in-memory-provisional-donation.repository';
 import { GiftogetherExceptions } from '../../filters/giftogether-exception';
